Add magic methods to retrieve relationships
E.g. calling `$model->items` will retrieve all rows from the 'items' table with
a relationship to the current model.

Calling `$model->items()` returns a query on the 'items' table with the join for
the relationship already applied, which can then be filtered further.

`$model->items` is equivalent to `$model->items()->fetch`.